Add files via upload
TASK - 2
1) Practiced basic to advance Linux commands.
2) Performed a "DOS" attack on the given web site by installing a tool "GoldenEye" from the Github.

To install GoldenEye from GitHub in your Kali Linux, you can follow these steps:

Open your terminal and use the following command to clone the GitHub repository:

git clone https://lnkd.in/gBSJe4yA

Move to the Goldeneye directory with this command:
cd GoldenEye

List the contents of the directory and run the tool with these commands:
ls
./goldeneye.py

The tool will ask for a URL to attack. You can also use the -h option to see the usage and options of the tool. For example, you can use the -s option to specify the number of concurrent sockets, or the -m option to choose the HTTP method (get, post, or random).

./goldeneye.py -h
./goldeneye.py https://www.example.com -s 1000 -m random